Refactor, add tests, and fix runtime issues
Refactoring:
- G6Format: constants made final, removed dead fields (g6_graph, bit_[]),
  unused imports dropped, duplicate graphsize() call removed, SETWD/SETBT
  made static, R() renamed to encodeChunks() with clarifying comment,
  string concatenation replaced with StringBuilder throughout
- Utils: getDegreeSum sum variable changed from int to double to avoid
  silent truncation of fractional results
- TotalGraph, VertexSemitotalGraph, EdgeSemitotalGraph,
  BarycentricSubdivisionGraph: removed unused `double dis` variable
- Union, CoronaProduct: standardised parameter names (fG/sG -> g0/g1)
  to match the other product action classes; removed unused GraphSum instance
- RandomGenerator: replaced recursive randomEdge() with iterative do/while
  loop to eliminate StackOverflowError risk on small graphs

Bug fixes:
- GraphSaveObject: replaced removed javax.xml.bind.DatatypeConverter with
  java.util.Base64 to fix compilation on Java 11+
- StreamingReport: removed time-windowed aggregation that caused zero output
  on bounded file sources (processing-time window never fired)

Tests (47 total, new):
- G6FormatTest: graphsize, stringToGraph, stringToGraphModel, roundtrip,
  createAdjMatrix, SIZELEN boundary values
- UtilsTest: choose, getMaxDegree, getDegreeSum, getMinNonPendentDegree,
  getBinaryPattern, getLaplacian
- BarycentricSubdivisionGraphTest: vertex/edge count formula |V'|=|V|+k|E|,
  |E'|=(k+1)|E| verified for k=0..3 on path and K3
- RandomGeneratorTest: correct vertex count, edge deduplication, iterative
  fix handles tiny graphs without stack overflow

Tooling:
- run.sh: build, test, stream, spectral commands with --add-opens flags
  required by Flink 1.6 on Java 9+
- README: updated with quick start, run.sh reference, entry point table,
  Java compatibility notes, project structure, and test instructions

+++ src/main/java/graphtea/extensions/Utils.java
@@ -43,9 +43,9 @@ public class Utils {
     }
 
     public static double getDegreeSum(GraphModel g, double alpha) {
-        int sum = 0;
+        double sum = 0;
         for(Vertex v: g) {
-            sum+=getDegreeSumOfVertex(g,alpha,v);
+            sum += getDegreeSumOfVertex(g, alpha, v);
         }
         return sum;
     }
